requirements.txt
certifi==2020.12.5
click==7.1.2
dataclasses==0.8
fastapi==0.63.0
h11==0.11.0
psycopg2-binary==2.8.6
pydantic==1.7.3
python-dotenv==0.15.0
SQLAlchemy==1.3.22
starlette==0.13.6
typing-extensions==3.7.4.3
uvicorn==0.13.2
